macOS port: first working build, config dialog confirmed
Port the project to macOS (Apple Silicon arm64) with no errors or
warnings — extension loads, registers in REAPER, and the surface
configuration dialog opens without crashing.

Core changes
------------
- Boost 1.39.0 → 1.91.0 (headers only):
  Drops all old-Boost patches (shared_ptr move-ctor, expired_slot
  incomplete type, auto_ptr / unary_function / binary_function removal).
  Modern Boost compiles cleanly under C++17 / libc++ on all platforms.

- SWELL binding on macOS (the crash fix):
  REAPER calls SWELL_dllMain with getfunc=NULL on macOS (load probe
  only), leaving every SWELL function pointer NULL → configFunc
  crashes with SIGSEGV on the first dialog call.
  Fix: compile swell-modstub.mm instead of swell-modstub-generic.cpp.
  Its global SwellAPPInitializer constructor binds SWELL via REAPER's
  NSApp delegate at dylib load time — the canonical macOS mechanism.

- APFS case-insensitivity fixes (repo root is on -I path):
  VERSION  → VERSION.txt  (collided with libc++ <version> header)
  Assert.h → McuAssert.h  (collided with system <assert.h>)

- Clang strictness fixes (GCC only warned):
  Removed redundant class qualification in csurf_mcu.h / Transport.h,
  narrowing cast via static_cast<unsigned char>, void*→int via
  intptr_t, address-of-temporary → named locals.

- Build system (CMakeLists.txt, APPLE branch):
  Force Apple Clang (Homebrew GCC breaks juceaide), scope -std=c++17
  to CXX files only (JUCE registers .c sources), LANGUAGE CXX on .mm
  stub, removed -include;cstddef hack (no longer needed).

- build_and_run.sh: cross-platform (macOS + Linux), --quick default,
  --release / --debug for full builds.

- fetch_deps.sh: Boost 1.91.0 URL, removed old patches/sed_inplace,
  OS-aware next-steps output.

- res.rc_mac_menu recreated (empty placeholder).
- dispatchNextMessageOnSystemQueue guarded to #if JUCE_LINUX.
- .gitignore: added /boost_1_91_0/.

Documentation
-------------
- Mac-Port-Changes.md: complete change log for the macOS port.
- AGENTS.md: Boost 1.91.0 references updated.
- MEMD.md: decisions, bugs, changelogs, patterns for all steps.

Tested on macOS 26.5.1 (Apple Silicon M4 Max), REAPER 7.76.

+++ MultiTrackMode.cpp
@@ -8,7 +8,7 @@
 #include "MultiTrackMeterBridge.h"
 #include "CCSManager.h"
 #include "csurf_mcu.h"
-#include "assert.h"
+#include "McuAssert.h"
 #include "stdio.h"
 #include "Display.h"
 #include "MultiTrackOptions.h"
@@ -177,7 +177,7 @@ void MultiTrackMode::updateAssignmentDisplay() {
   int nr;
   MediaTrack *pMT = Tracks::instance()->getSelectedSingleTrack();
   if (pMT) {
-    nr = (int)GetSetMediaTrackInfo(pMT, "IP_TRACKNUMBER", NULL);
+    nr = (int)(intptr_t)GetSetMediaTrackInfo(pMT, "IP_TRACKNUMBER", NULL);
 
     char display[2];
     display[0] = '0' + ((nr / 10) % 10);
